requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
djangorestframework==3.15.2
python_dotenv==1.0.1
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
django-cors-headers==4.6.0
drf_yasg==1.21.8